chore: PostImageContainer 스토리북 추가 (navigate 사용시 Router 오류 )
